requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.3.0
marshmallow==3.20.1
